increase RAM to 2G to be able to boot
With the default RAM 1G, after the default Container Linux iso image
got unpacked, the Qemu VM fails to boot with an error message "Failed
to switch root". Thus it's not possible at all to proceed installation.
In these days, it's fairly reasonable to set the default RAM to 2G
in most cases. So let's increase the default RAM to 2G.